+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 **New features**
 - Address lookup source: allows to choose a source different from forecast source for address lookup of current location. Docs were updated to use the terms “address lookup” where it was previously called “reverse geocoding”
 
+**Weather sources**
+- Android Geocoder: added as possible Address lookup source on devices supporting it
+
 
 # Version 5.4.2 (2025-02-08)